chore: migrate Page tests
